simulate gravity on cpu, first go at cuda

+++ src/math/units.h
@@ -3,7 +3,7 @@
 
 typedef unsigned long long int biguint;
 typedef signed long long int bigint;
-typedef long double real;
+typedef double real;
 
 typedef real seconds;
 typedef real kilograms;
